Updated position and weld snapping numbers to be true to the minecraft pixel grid Changed rotation angle from Int to Double and added 22.5 degrees as a rotation snapping option

+++ src/main/java/com/enxv/aeronauticsstructuretool/ClientToolState.java
@@ -161,7 +161,7 @@ public final class ClientToolState {
         persist();
     }
 
-    public static int getRotationStep() {
+    public static double getRotationStep() {
         return PREFERENCES.rotationStep();
     }
 
@@ -179,7 +179,7 @@ public final class ClientToolState {
         persist();
     }
 
-    public static int getRotationDegrees() {
+    public static double getRotationDegrees() {
         return PREFERENCES.rotationDegrees();
     }
